simplifying top navigations
