--- App/navigation/HomeNavigator.js
@@ -1,23 +1,0 @@
-import React from 'react';
-import { createNativeStackNavigator } from '@react-navigation/native-stack';
-import HomeScreen from '../screens/HomeScreen';
-
-const Stack = createNativeStackNavigator();
-const HomeNavigator = ({ navigation }) => {
-  return (
-    <Stack.Navigator screenOptions={{contentStyle: {backgroundColor: '#F5F5F5'}}}>
-      <Stack.Screen 
-        options={screenOptionsWithBackButton('Eon English')} 
-        name="Home" 
-        component={HomeScreen}
-      />
-    </Stack.Navigator>
-  );
-};
-
-const screenOptionsWithBackButton = (title) => ({ navigation }) => ({
-  title: title,
-  headerTitleAlign: 'center'
-});
-
-export default HomeNavigator;
